Various PIPS OPF changes.

+++ pypower/makeAy.py
@@ -15,8 +15,8 @@
 # along with PYPOWER. If not, see <http://www.gnu.org/licenses/>.
 
 from numpy import array, diff, any, zeros, r_, flatnonzero as find
-from scipy.sparse import csr_matrix as sparse
-#from scipy.sparse import lil_matrix as sparse
+#from scipy.sparse import csr_matrix as sparse
+from scipy.sparse import lil_matrix as sparse
 
 from idx_cost import MODEL, PW_LINEAR, NCOST, COST
 
@@ -66,14 +66,14 @@ def makeAy(baseMVA, ng, gencost, pgbas, qgbas, ybas):
     ## same order as the compressed column sparse format used by matlab
     ## this should be the quickest.
 
-    m = sum(gencost[iycost, NCOST])  ## total number of cost points
+    m = sum( gencost[iycost, NCOST].astype(int) )  ## total number of cost points
     Ay = sparse((m - ny, ybas + ny - 1))
     by = array([])
     ## First fill the Pg or Qg coefficients (since their columns come first)
     ## and the rhs
     k = 0
     for i in iycost:
-        ns = gencost[i, NCOST]              ## # of cost points segments = ns-1
+        ns = gencost[i, NCOST].astype(int)         ## # of cost points segments = ns-1
         p = gencost[i, COST:COST + 2 * ns - 1:2] / baseMVA
         c = gencost[i, COST + 1:COST + 2 * ns:2]
         m = diff(c) / diff(p)               ## slopes for Pg (or Qg)
@@ -96,7 +96,7 @@ def makeAy(baseMVA, ng, gencost, pgbas, qgbas, ybas):
     k = 0
     j = 0
     for i in iycost:
-        ns = gencost[i, NCOST]
+        ns = gencost[i, NCOST].astype(int)
         ## FIXME: Bug in SciPy 0.7.2 prevents setting with a sequence
 #        Ay[k:k + ns - 1, ybas + j - 1] = -ones(ns - 1)
         for kk in range(k, k + ns - 1):
